# lib/api-zod/src/generated/types/progressRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProgressRowStatus } from "./progressRowStatus";

export interface ProgressRow {
  id: number;
  userId: number;
  slokaId: string;
  status: ProgressRowStatus;
  updatedAt: Date;
}